GH #2 - Check given input values for sanity
This commit brings sanity checking for configuration given by receptor
to the plugin for a run to make the plugin more robust. If a a
configuration value does not pass the checks, a warning is logged and
default value is used instead.

This also checks if hostnames contain a comma, such hostnames not sent
over to Satellite are marked as failed straight away.